Revert "## Fix zoho sync precedance"
